Move category delete dialog path to qs
