--- saleor/static/dashboard-next/categories/views/CategoryDetails/urls.ts
@@ -1,7 +1,0 @@
-import * as urlJoin from "url-join";
-import { categoryPath } from "../../urls";
-
-export const categoryDeletePath = (id: string) =>
-  urlJoin(categoryPath(id), "delete");
-export const categoryDeleteUrl = (id: string) =>
-  categoryDeletePath(encodeURIComponent(id));
